+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 -----
 - Added a new Domain class for Memsource REST transition.
 - Added a new Language class for Memsource REST transition.
+- Added a new TermBase class for Memsource REST transition.
 
 [0.5.1] - 2020-06-19
 ====================